fix: Re-order methods in PageContextService and rename files to match naming scheme
I didn't followed the naming scheme previously. This commit fixes that. Additionally, I re-ordered the methods inside the PageContextService and removed the suppression comment.
