+++ deb/openmediavault/workbench/src/app/core/components/intuition/form/form.component.ts
@@ -42,7 +42,7 @@ import { Unsubscribe } from '~/app/decorators';
 import { format, formatDeep } from '~/app/functions.helper';
 import { CustomValidators } from '~/app/shared/forms/custom-validators';
 import { ConstraintService } from '~/app/shared/services/constraint.service';
-import { PageContextService } from '~/app/shared/services/page-context-service';
+import { PageContextService } from '~/app/shared/services/page-context.service';
 
 let nextUniqueId = 0;
